Device: Cast init_dev->local to uint64_t, should fix the warning.

+++ src/device.c
@@ -362,7 +362,7 @@ device_add_common(const device_t *dev, void *p, void *params, int inst)
         if ((devices[c] == dev) && (device_state[c].local == init_dev->local) &&
             (device_state[c].inst == inst)) {
             warning("DEVICE: Device \"%s\", local 0x%016" PRIX64 ", inst %i already exists!\n",
-                       init_dev->name, init_dev->local,
+                       init_dev->name, (uint64_t) init_dev->local,
                        inst);
             return (NULL);
         }
